chore(orchestrator): moves all the /next components
Now they are located inside the /components directory
