+++ plugins/orchestrator/src/components/CreateWorkflowPage.tsx
@@ -16,17 +16,14 @@ import Button from '@material-ui/core/Button';
 
 import { WORKFLOW_TITLE } from '@janus-idp/backstage-plugin-orchestrator-common';
 
-import { orchestratorApiRef } from '../../api';
-import {
-  editWorkflowRouteRef,
-  workflowDefinitionsRouteRef,
-} from '../../routes';
-import { BaseOrchestratorPage } from '../next/BaseOrchestratorPage';
-import { WorkflowEditor } from '../WorkflowEditor';
+import { orchestratorApiRef } from '../api';
+import { editWorkflowRouteRef, workflowDefinitionsRouteRef } from '../routes';
+import { BaseOrchestratorPage } from './BaseOrchestratorPage';
+import { WorkflowEditor } from './WorkflowEditor';
 import {
   EditorViewKind,
   WorkflowEditorRef,
-} from '../WorkflowEditor/WorkflowEditor';
+} from './WorkflowEditor/WorkflowEditor';
 
 export const CreateWorkflowPage = () => {
   const { format } = useParams();
