--- plugins/orchestrator/src/components/CreateWorkflowPage/index.ts
@@ -1,1 +1,0 @@
-export { CreateWorkflowPage } from './CreateWorkflowPage';
