+++ plugins/orchestrator/src/components/ExecuteWorkflowPage/ExecuteWorkflowPage.tsx
@@ -25,7 +25,7 @@ import {
   workflowInstanceRouteRef,
 } from '../../routes';
 import { getErrorObject } from '../../utils/errorUtils';
-import { BaseOrchestratorPage } from '../next/BaseOrchestratorPage';
+import { BaseOrchestratorPage } from '../BaseOrchestratorPage';
 import JsonTextAreaForm from './JsonTextAreaForm';
 import StepperForm from './StepperForm';
