--- plugins/orchestrator/src/components/NewWorkflowViewerPage/index.ts
@@ -1,1 +1,0 @@
-export { NewWorkflowViewerPage } from './NewWorkflowViewerPage';
