+++ plugins/orchestrator/src/components/OrchestratorPage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { TabbedLayout } from '@backstage/core-components';
 
-import { workflowInstancesRouteRef } from '../../routes';
+import { workflowInstancesRouteRef } from '../routes';
 import { BaseOrchestratorPage } from './BaseOrchestratorPage';
 import { WorkflowRunsTabContent } from './WorkflowRunsTabContent';
 import { WorkflowsTabContent } from './WorkflowsTabContent';
