+++ plugins/orchestrator/src/components/WorkflowDefinitionViewerPage/WorkflowDefinitionViewerPage.tsx
@@ -21,8 +21,8 @@ import {
   executeWorkflowRouteRef,
   workflowDefinitionsRouteRef,
 } from '../../routes';
-import { BaseOrchestratorPage } from '../next/BaseOrchestratorPage';
-import EditWorkflowDialog from '../WorkflowDialog/EditWorkflowDialog';
+import { BaseOrchestratorPage } from '../BaseOrchestratorPage';
+import EditWorkflowDialog from '../EditWorkflowDialog';
 import { EditorViewKind, WorkflowEditor } from '../WorkflowEditor';
 import WorkflowDefinitionDetailsCard from './WorkflowDefinitionDetailsCard';
