--- plugins/orchestrator/src/components/WorkflowDialog/index.ts
@@ -1,1 +1,0 @@
-export { WorkflowDialog } from './WorkflowDialog';
