+++ plugins/orchestrator/src/components/WorkflowInstancePageContent.tsx
@@ -12,13 +12,10 @@ import {
   ProcessInstanceStateValues,
 } from '@janus-idp/backstage-plugin-orchestrator-common';
 
-import { VALUE_UNAVAILABLE } from '../../constants';
-import {
-  executeWorkflowRouteRef,
-  workflowInstanceRouteRef,
-} from '../../routes';
-import { capitalize } from '../../utils/StringUtils';
-import { EditorViewKind, WorkflowEditor } from '../WorkflowEditor';
+import { VALUE_UNAVAILABLE } from '../constants';
+import { executeWorkflowRouteRef, workflowInstanceRouteRef } from '../routes';
+import { capitalize } from '../utils/StringUtils';
+import { EditorViewKind, WorkflowEditor } from './WorkflowEditor';
 import { WorkflowInstanceStatusIndicator } from './WorkflowInstanceStatusIndicator';
 import { WorkflowProgress } from './WorkflowProgress';
 import { WorkflowRunDetail, WorkflowSuggestion } from './WorkflowRunDetail';
